Refactored cluster provider, documented API
