+++ src/extension.api.ts
@@ -1,0 +1,6 @@
+import * as clusterproviderregistry from './components/clusterprovider/clusterproviderregistry';
+
+export interface ExtensionAPI {
+    readonly apiVersion: string;
+    readonly clusterProviderRegistry: clusterproviderregistry.ClusterProviderRegistry;
+}
